interactive list functionality
